requirements.txt
affine==2.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
cogeo-mosaic==8.2.0
color-operations==0.2.0
cramjam==2.11.0
fastapi==0.124.0
geojson-pydantic==2.1.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.3
morecantile==6.2.0
numexpr==2.14.1
numpy==2.3.5
opentelemetry-api==1.39.0
opentelemetry-exporter-otlp==1.39.0
opentelemetry-exporter-otlp-proto-common==1.39.0
opentelemetry-exporter-otlp-proto-grpc==1.39.0
opentelemetry-exporter-otlp-proto-http==1.39.0
opentelemetry-instrumentation==0.60b0
opentelemetry-instrumentation-asgi==0.60b0
opentelemetry-instrumentation-fastapi==0.60b0
opentelemetry-instrumentation-logging==0.60b0
opentelemetry-proto==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-semantic-conventions==0.60b0
opentelemetry-util-http==0.60b0
packaging==25.0
protobuf==6.33.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyparsing==3.2.5
pyproj==3.7.2
pystac==1.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
rasterio==1.4.3
requests==2.32.5
rio-cogeo==5.4.2
rio-stac==0.12.0
rio-tiler==7.9.2
shapely==2.1.2
simplejson==3.20.2
six==1.17.0
starlette==0.50.0
starlette-cramjam==0.5.1
supermorecado==0.1.2
titiler-application==0.26.0
titiler-core==0.26.0
titiler-extensions==0.26.0
titiler-mosaic==0.26.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.1
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
zipp==3.23.0
